move project stuff into a project specific folder
