+++ frontend/viewer/src/project/browse/BrowseView.svelte
